icontrol/requirements.txt
asgiref==3.2.10
Django==2.2.16
django-admin-sortable2==0.7.7
django-cleanup==5.1.0
django-modeladmin-reorder==0.3.1
django-mptt==0.11.0
django-mptt-admin==2.0.0
django-select2==7.4.2
django-smart-selects==1.5.4
gunicorn==20.0.4
Pillow==7.2.0
pytz==2020.1
sqlparse==0.4.1
https://github.com/darklow/django-suit/tarball/v2
git+https://github.com/digi604/django-smart-selects.git@js-unlinting-fixes
